cleaned up artillery python wrapper methods
